feat: main.pyが大きくなってきたためStripeAPIを分離、DB接続やセッション管理のためのdatabase.py作成
